# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: MiniApp API
  version: 0.1.0
  description: API contract shared by the UniApp client and FastAPI service.
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: Local development
paths:
  /health:
    get:
      operationId: getHealth
      summary: Check API health
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /items:
    get:
      operationId: listItems
      summary: List items
      responses:
        "200":
          description: Items ordered from newest to oldest
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemRead"
    post:
      operationId: createItem
      summary: Create an item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemCreate"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemRead"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /items/{item_id}:
    parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
    put:
      operationId: updateItem
      summary: Update an item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemRead"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      operationId: deleteItem
      summary: Delete an item
      responses:
        "204":
          description: Item deleted
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
components:
  schemas:
    ErrorDetail:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    HealthResponse:
      type: object
      required: [status, service]
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: MiniApp API
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
    ItemCreate:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 80
          pattern: ".*\\S.*"
    ItemUpdate:
      type: object
      additionalProperties: false
      required: [completed]
      properties:
        completed:
          type: boolean
    ItemRead:
      type: object
      required: [id, title, completed, created_at]
      properties:
        id:
          type: integer
        title:
          type: string
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time